Added deactivation parameter. Added dynamically populated select.
Added deactivation  parameter for all the component types: a string
containing a function body that must return a boolean value in order to
decide whether the parameter should be disabled in the form.

Added the possibility to dynamically populate a select: if the
parameter has a “isDynamic” property set to the boolean value true, the
directive will search for a property named: getDynamicValues with this
format:
“var data = [{\"name\": \"ciao\",\"value\": 1}]; cb(err, data));”. Here
the data variable is an array containing object with two properties:
name and value. cb is a callback that has to be invoked with two
parameters: err containing the error, and data containing the data.